refactor: split UI components and events
